+++ bibfor/algeline/bmocca.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -54,7 +54,7 @@ subroutine bmocca(umoy, geom, cf0, mcf0, fsvr, &
 #include "asterfort/pbflui.h"
 #include "asterfort/wkvect.h"
     real(kind=8) :: umoy, geom(9), cf0, mcf0, fsvr(7)
-    integer :: nbm, vicoq(nbm)
+    integer(kind=8) :: nbm, vicoq(nbm)
     real(kind=8) :: torco(4, nbm), tcoef(10, nbm), s1, s2
     complex(kind=8) :: b(nbm, nbm)
 !
@@ -64,9 +64,9 @@ subroutine bmocca(umoy, geom, cf0, mcf0, fsvr, &
 !-----------------------------------------------------------------------
 !
 !-----------------------------------------------------------------------
-    integer :: icoq, iligne, imod, iysol, iysol1, iysol2, jcoq
-    integer :: jligne, jmod, k1i, k1j, k2i, k2j, ki
-    integer :: kj
+    integer(kind=8) :: icoq, iligne, imod, iysol, iysol1, iysol2, jcoq
+    integer(kind=8) :: jligne, jmod, k1i, k1j, k2i, k2j, ki
+    integer(kind=8) :: kj
     real(kind=8) :: hmoy, r1, r2, rhof, rk1i, rk1j, rk2i
     real(kind=8) :: rk2j, rki, rkj, rmoy, thet1i, thet1j, thet2i
     real(kind=8) :: thet2j, thetai, thetaj
