+++ bibfor/algeline/calcdy.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -38,7 +38,7 @@ subroutine calcdy(mu, k, f0, devg, devgii, &
 ! --- : DELTA  : DELTA LAMBDA INITIAL ----------------------------------
 ! OUT : DY     : INCREMENTS DE L'ITERATION COURANTE --------------------
 ! ======================================================================
-    integer :: ii, ndt, ndi
+    integer(kind=8) :: ii, ndt, ndi
     real(kind=8) :: ddelta, dgamp, dsn(6), dinv, devp, mun, deux, trois
 ! ======================================================================
 ! --- INITIALISATION DE PARAMETRES -------------------------------------
