+++ bibfor/algeline/calcg.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -37,7 +37,7 @@ subroutine calcg(dfds, vecn, g, devg, traceg, &
 ! --- : TRACEG : PREMIER INVARIANT DE G --------------------------------
 ! --- : DEVGII : NORME DU DEVIATEUR ------------------------------------
 ! ======================================================================
-    integer :: ii, ndt, ndi
+    integer(kind=8) :: ii, ndt, ndi
     real(kind=8) :: fact1
     blas_int :: b_incx, b_incy, b_n
 ! ======================================================================
