+++ bibfor/algeline/calcj0.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -32,10 +32,10 @@ subroutine calcj0(t, sigpri, valp)
     real(kind=8) :: valp(3), jacaux(3)
     real(kind=8) :: tol, toldyn
     real(kind=8) :: sigpri
-    integer :: nbvec, nperm
-    integer :: i, itype, iordre
+    integer(kind=8) :: nbvec, nperm
+    integer(kind=8) :: i, itype, iordre
 !-----------------------------------------------------------------------
-    integer :: nitjac
+    integer(kind=8) :: nitjac
 !-----------------------------------------------------------------------
     data nperm, tol, toldyn/12, 1.d-10, 1.d-2/
 ! ----------------------------------------------------------------------
