+++ bibfor/algeline/chcoma.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -63,7 +63,7 @@ subroutine chcoma(tablez, nomaou)
 !                                 ROUTINE.
 !.========================= DEBUT DES DECLARATIONS ====================
 ! -----  VARIABLES LOCALES
-    integer :: ngm, ibid, iret, idcode, dimcoo, nbno, jcoor, idcoor, ino
+    integer(kind=8) :: ngm, ibid, iret, idcode, dimcoo, nbno, jcoor, idcoor, ino
     real(kind=8) :: r8b, p(2, 2), alpha, yg, zg, yabs, zabs
     complex(kind=8) :: c16b
     character(len=8) :: k8b, typobj
