+++ bibfor/algeline/crsvmf.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -26,7 +26,7 @@ subroutine crsvmf(motfac, solveu, istop, nprec, &
 #include "asterfort/jemarq.h"
 #include "asterfort/jeveuo.h"
 !
-    integer :: istop, nprec
+    integer(kind=8) :: istop, nprec
     real(kind=8) :: epsmat
     character(len=3) :: mixpre, kellag
     character(len=8) :: kxfem
@@ -47,9 +47,9 @@ subroutine crsvmf(motfac, solveu, istop, nprec, &
 !
 !
 !
-    integer ::    ibid
+    integer(kind=8) ::    ibid
     character(len=8) :: renum
-    integer, pointer :: slvi(:) => null()
+    integer(kind=8), pointer :: slvi(:) => null()
     real(kind=8), pointer :: slvr(:) => null()
     character(len=24), pointer :: slvk(:) => null()
 !
