+++ bibfor/algeline/detmat.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -32,7 +32,7 @@ subroutine detmat()
 #include "asterfort/detrsd.h"
 #include "asterfort/jeexin.h"
 #include "asterfort/jelstc.h"
-    integer :: nbmat, i, ier
+    integer(kind=8) :: nbmat, i, ier
     character(len=19) :: matass
     character(len=24) :: lirefa(100)
 !-----------------------------------------------------------------------
