+++ bibfor/algeline/factorial.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -24,8 +24,8 @@ function factorial(n)
 #include "asterfort/assert.h"
 !
 !
-    integer, intent(in) :: n
-    integer             :: factorial
+    integer(kind=8), intent(in) :: n
+    integer(kind=8)             :: factorial
 !
 ! --------------------------------------------------------------------------------------------------
 !
